requirements.txt
fastapi
uvicorn
pymongo
pymongo[srv]
passlib
bcrypt
python-jose
python-multipart